# src/main/resources/openapi/api.yaml
openapi: 3.0.1
info:
  title: Homni Feature Toggle API
  version: '2.0.0'
  description: Per-project feature toggle management with RBAC and API key authentication.
servers:
  - url: 'http://localhost:8080'
    description: local
security:
  - apiKey: []
  - oidc: []
...
paths:
...
  /projects/{projectId}/toggles:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Feature Toggles]
      operationId: listToggles
      summary: List feature toggles in a project
      description: Requires READ_TOGGLES permission (READER+).
      parameters:
        - name: enabled
          in: query
          required: false
          schema:
            type: boolean
        - name: environment
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureToggleListResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'

    post:
      tags: [Feature Toggles]
      operationId: createToggle
      summary: Create a new feature toggle
      description: Requires WRITE_TOGGLES permission (EDITOR+). Project must not be archived.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeatureToggleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureToggleSingleResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'
  /projects/{projectId}/toggles/{toggleId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: toggleId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Feature Toggles]
      operationId: getToggle
      summary: Get toggle by ID
      description: Requires READ_TOGGLES permission (READER+).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureToggleSingleResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

    patch:
      tags: [Feature Toggles]
      operationId: updateToggle
      summary: Update a feature toggle
      description: Partial update. Requires WRITE_TOGGLES permission (EDITOR+).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFeatureToggleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureToggleSingleResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

    delete:
      tags: [Feature Toggles]
      operationId: deleteToggle
      summary: Delete a feature toggle
      description: Requires WRITE_TOGGLES permission (EDITOR+).
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

  # ========== Environments ==========
  /projects/{projectId}/environments:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Environments]
      operationId: listEnvironments
      summary: List environments in a project
      description: Requires READ_TOGGLES permission (READER+).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentListResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'

    post:
      tags: [Environments]
      operationId: createEnvironment
      summary: Create a new environment
      description: Requires WRITE_TOGGLES permission (EDITOR+). Project must not be archived.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnvironmentRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentSingleResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'
  /projects/{projectId}/environments/{environmentId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Environments]
      operationId: deleteEnvironment
      summary: Delete an environment
      description: Requires WRITE_TOGGLES permission (EDITOR+). Cannot delete if in use by toggles.
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'

  # ========== API Keys ==========
  /projects/{projectId}/api-keys:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [API Keys]
      operationId: listApiKeys
      summary: List API keys (tokens are masked)
      description: Requires MANAGE_MEMBERS permission (ADMIN).
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyListResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'

    post:
      tags: [API Keys]
      operationId: issueApiKey
      summary: Issue a new read-only API key
      description: Requires MANAGE_MEMBERS permission (ADMIN). API keys always grant READER role (read-only). The raw token is returned only once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueApiKeyRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreatedSingleResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  /projects/{projectId}/api-keys/{apiKeyId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: apiKeyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [API Keys]
      operationId: revokeApiKey
      summary: Revoke an API key
      description: Requires MANAGE_MEMBERS permission (ADMIN).
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

  # ========== Members ==========
...
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    '204':
      description: No content
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '409':
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ResponseMeta:
      type: object
      required: [timestamp]
      properties:
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # ---------- Project ----------
...
    CreateFeatureToggleRequest:
      type: object
      required: [name, environments]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: dark-mode
        description:
          type: string
          maxLength: 1000
        environments:
          type: array
          items:
            type: string
          example: [DEV, PROD]
    UpdateFeatureToggleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        environments:
          type: array
          items:
            type: string
        enabled:
          type: boolean
    FeatureToggle:
      type: object
      required: [id, projectId, name, enabled, environments, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        environments:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    FeatureToggleSingleResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          $ref: '#/components/schemas/FeatureToggle'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    FeatureToggleListResponse:
      type: object
      required: [payload, pagination, meta]
      properties:
        payload:
          type: array
          items:
            $ref: '#/components/schemas/FeatureToggle'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # ---------- Environment ----------
    CreateEnvironmentRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: STAGING
    Environment:
      type: object
      required: [id, projectId, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    EnvironmentSingleResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          $ref: '#/components/schemas/Environment'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    EnvironmentListResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          type: array
          items:
            $ref: '#/components/schemas/Environment'
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # ---------- API Key ----------
    IssueApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: ci-cd-pipeline
        expiresAt:
          type: string
          format: date-time
    ApiKey:
      type: object
      required: [id, projectId, name, role, maskedToken, active, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, EDITOR, READER]
        maskedToken:
          type: string
          example: 'hft_****abcd'
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
    ApiKeyCreated:
      type: object
      required: [id, name, role, rawToken, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, EDITOR, READER]
        rawToken:
          type: string
          example: 'hft_a1b2c3d4e5f6g7h8i9j0'
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
    ApiKeyListResponse:
      type: object
      required: [payload, pagination, meta]
      properties:
        payload:
          type: array
          items:
            $ref: '#/components/schemas/ApiKey'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ApiKeyCreatedSingleResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          $ref: '#/components/schemas/ApiKeyCreated'
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # ---------- Members ----------
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    oidc:
      type: http
      scheme: bearer
      bearerFormat: JWT